% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_w}
\alias{calc_w}
\title{Truncation correction ratio phi(w) / Phi(w)}
\usage{
calc_w(sigma, nu)
}
\arguments{
\item{sigma}{Approximate CV parameter.}

\item{nu}{Skewness parameter (nu != 0).}
}
\value{
Scalar ratio dnorm(1/(sigma|nu|)) / pnorm(1/(sigma|nu|)).
}
\description{
Used in the score and Hessian components for sigma and nu when nu != 0.
}
\keyword{internal}
