% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdpde_bccg.R
\name{calculate_k}
\alias{calculate_k}
\title{Estimate the K matrix of the sandwich covariance}
\usage{
calculate_k(theta, y, X_mu, X_sigma, X_nu, alpha, mu_link, sigma_link, nu_link)
}
\description{
K = (1/n) sum psi_i psi_i' - xi xi', where psi_i = u(y_i) f(y_i)^alpha.
}
\keyword{internal}
