% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdpde_bccg.R
\name{coef.mdpde_bccg}
\alias{coef.mdpde_bccg}
\title{Extract coefficients from an mdpde_bccg object}
\usage{
\method{coef}{mdpde_bccg}(object, parameter = c("all", "mu", "sigma", "nu"), ...)
}
\arguments{
\item{object}{An object of class \code{"mdpde_bccg"}.}

\item{parameter}{Which sub-model: \code{"all"} (default), \code{"mu"}, \code{"sigma"}, or
\code{"nu"}.}

\item{...}{Ignored.}
}
\description{
Extract coefficients from an mdpde_bccg object
}
