% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{link_derivative}
\alias{link_derivative}
\title{First derivative of the inverse link (d mu / d eta)}
\usage{
link_derivative(theta, link)
}
\arguments{
\item{theta}{Parameter vector (on the response scale).}

\item{link}{Character string.}
}
\value{
Derivative vector.
}
\description{
First derivative of the inverse link (d mu / d eta)
}
\keyword{internal}
