% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{link_inverse}
\alias{link_inverse}
\title{Inverse link function}
\usage{
link_inverse(eta, link)
}
\arguments{
\item{eta}{Linear predictor vector.}

\item{link}{Character string: \code{"log"}, \code{"identity"}, or \code{"sqrt"}.}
}
\value{
Transformed parameter vector.
}
\description{
Inverse link function
}
\keyword{internal}
