% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{link_second_derivative}
\alias{link_second_derivative}
\title{Second derivative of the inverse link (d^2 mu / d eta^2)}
\usage{
link_second_derivative(theta, link)
}
\arguments{
\item{theta}{Parameter vector (on the response scale).}

\item{link}{Character string.}
}
\value{
Second derivative vector.
}
\description{
Second derivative of the inverse link (d^2 mu / d eta^2)
}
\keyword{internal}
