% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{observed_information_matrix}
\alias{observed_information_matrix}
\title{Observed Fisher information matrix for a single observation}
\usage{
observed_information_matrix(
  y_i,
  mu_i,
  sigma_i,
  nu_i,
  x_mu_i,
  x_sigma_i,
  x_nu_i,
  mu_link,
  sigma_link,
  nu_link
)
}
\description{
Observed Fisher information matrix for a single observation
}
\keyword{internal}
