% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdpde_bccg.R
\name{print.mdpde_bccg}
\alias{print.mdpde_bccg}
\title{Print an mdpde_bccg object}
\usage{
\method{print}{mdpde_bccg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mdpde_bccg"}.}

\item{...}{Ignored.}
}
\description{
Print an mdpde_bccg object
}
