% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{score_mu}
\alias{score_mu}
\title{Score function for mu (individual observation)}
\usage{
score_mu(y, mu, sigma, nu)
}
\arguments{
\item{y, mu, sigma, nu}{BCCG parameters (scalars).}
}
\value{
Scalar dl/dmu.
}
\description{
Score function for mu (individual observation)
}
\keyword{internal}
