% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{score_sigma}
\alias{score_sigma}
\title{Score function for sigma (individual observation)}
\usage{
score_sigma(y, mu, sigma, nu)
}
\arguments{
\item{y, mu, sigma, nu}{BCCG parameters (scalars).}
}
\value{
Scalar dl/dsigma.
}
\description{
Score function for sigma (individual observation)
}
\keyword{internal}
